refactor: move types in their own file
